XIVY-3255 move download impl to own package
